10.0.2

+++ scripts/_promptListeners.mjs
@@ -1,4 +1,6 @@
-export function setHooks_promptListeners(){
+import { MODULE } from "./settings.mjs";
+
+export function setHooks_promptListeners() {
 
     Hooks.on("renderChatLog", (_, html) => {
         html[0].addEventListener("click", (event) => {
@@ -12,36 +14,37 @@ export function setHooks_promptListeners(){
     });
 }
 
-async function clickConcentrationPrompt(event){
+async function clickConcentrationPrompt(event) {
     // get the target of the mouse click.
     let button = event.target?.closest("button");
     if ( !button ) button = event.target?.closest("[name='render-item-sheet']");
     if ( !button ) return;
 
-    if ( button.name === "saving-throw-button" ){
-        const {dc, saveType, actorUuid} = button.dataset;
+    if ( button.name === "saving-throw-button" ) {
+        const { dc, saveType, actorUuid } = button.dataset;
         const actor = fromUuidSync(actorUuid);
-        return actor.rollConcentrationSave(saveType, {targetValue: dc});
+        return actor.rollConcentrationSave(saveType, { targetValue: dc });
     }
-    else if ( button.name === "delete-concentration-button" ){
-        const {effectUuid} = button.dataset;
+    else if ( button.name === "delete-concentration-button" ) {
+        const { effectUuid } = button.dataset;
         const effect = fromUuidSync(effectUuid);
         return deleteDialog(effect, event);
     }
-    else if ( button.name === "render-item-sheet" ){
-        const {itemUuid} = button.dataset;
+    else if ( button.name === "render-item-sheet" ) {
+        const { itemUuid } = button.dataset;
         const item = fromUuidSync(itemUuid);
         return item.sheet.render(true);
     }
 }
 
-async function deleteDialog(effect, event){
-    if ( event.shiftKey ){
+async function deleteDialog(effect, event) {
+    if ( !effect ) return;
+    if ( event.shiftKey ) {
         return effect.delete();
     }
-    const name = effect.getFlag("concentrationnotifier", "data.itemData.name");
-    const title = game.i18n.format("CN.DELETE_DIALOG_TITLE", {name});
-    const content = game.i18n.format("CN.DELETE_DIALOG_TEXT", {name});
+    const name = effect.getFlag(MODULE, "data.itemData.name");
+    const title = game.i18n.format("CN.DELETE_DIALOG_TITLE", { name });
+    const content = game.i18n.format("CN.DELETE_DIALOG_TEXT", { name });
     const labelYes = game.i18n.localize("Yes");
     const labelNo = game.i18n.localize("No");
     new Dialog({
